Refactor routes

+++ app/Http/Controllers/PlaylistController.php
@@ -10,6 +10,7 @@ use App\Models\Playlist;
 use App\Models\User;
 use App\Models\Video;
 use Exception;
+use http\Env\Response;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 
@@ -72,6 +73,10 @@ class PlaylistController extends Controller
      */
     public function update(PlaylistUpdate $request, Playlist $playlist)
     {
+        if (!request()->is('api/admin/*') && $playlist->user_id != auth('api')->id()){
+            abort(403, 'Access denied');
+        }
+
         $playlist->name = $request->get('name');
 
         $playlist->owner()->associate(Auth::user());
@@ -89,8 +94,14 @@ class PlaylistController extends Controller
      */
     public function destroy(Playlist $playlist)
     {
+        if (!request()->is('api/admin/*') && $playlist->user_id != auth('api')->id()){
+            abort(403, 'Access denied');
+        }
+
         $playlist->videos()->detach();
         $playlist->delete();
+
+        return response()->json(['message' => 'ok']);
     }
 
     /**
@@ -99,10 +110,8 @@ class PlaylistController extends Controller
      */
     public function add(Playlist $playlist, Video $video){
 
-        if($playlist->owner->id != Auth::user()->id){
-            return response()->json([
-               'message' => 'general.not_found'
-            ], 404);
+        if($playlist->user_id != Auth::user()->id){
+            abort(404, 'general.not_found');
         }
 
         $playlist->videos()->syncWithoutDetaching($video);
@@ -132,7 +141,9 @@ class PlaylistController extends Controller
     }
 
     public function bulkAdd(Request $request){
-        $playlists = Playlist::whereIn('id', $request->get('playlists',[]))->get();
+        $playlists = Playlist::where('user_id', auth('api')->id())
+            ->whereIn('id', $request->get('playlists',[]))
+            ->get();
         $videos = Video::whereIn('id', $request->get('videos', []))->get();
 
         $playlists->map(function($playlist) use($videos) {
